// admin/src/common/api/generated/api.ts
export interface components {
    schemas: {
// ...
        AuthMe: {
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /** Username */
            username: string;
            /** Email */
            email: string;
            /** Display Name */
            display_name: string;
            /** Avatar Url */
            avatar_url?: string | null;
            status: components["schemas"]["UserStatus"];
            /**
             * Roles
             * @default []
             */
            roles: string[];
            /**
             * Capabilities
             * @default []
             */
            capabilities: string[];
        };
// ...
        LoginRequest: {
            /** Identifier */
            identifier: string;
            /** Password */
            password: string;
        };
// ...
        RegisterRequest: {
            /** Username */
            username: string;
            /** Email */
            email: string;
            /** Password */
            password: string;
            /** Display Name */
            display_name?: string | null;
        };
// ...
        TokenPair: {
            /** Access Token */
            access_token: string;
            /** Refresh Token */
            refresh_token: string;
            /**
             * Token Type
             * @default bearer
             * @constant
             */
            token_type: "bearer";
            /** Expires In */
            expires_in: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}